requirements.txt
click==6.7
dj-database-url==0.4.2
Django==2.0.2
django-heroku==0.2.0
djangorestframework==3.7.7
djangorestframework-gis==0.12
Flask==0.12.2
geos==0.2.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
MarkupSafe==1.0
numpy==1.14.0
pandas==0.22.0
postgis==1.0.4
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
Werkzeug==0.14.1
whitenoise==3.3.1
django-rest-swagger==2.1.2